refactor: rename ProfileHeader container to ProfileEdit
