+++ src/screens/ProfileEdit/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { View } from 'react-native';
 
-import { ProfileHeader } from '../../containers';
+import { ProfileEdit as ProfileEditContainer } from '../../containers';
 
 import { ScreenContainer } from '../';
 
@@ -9,7 +9,7 @@ const ProfileEdit = () => {
   return (
     <ScreenContainer>
       <View style={{ marginTop: 50 }} />
-      <ProfileHeader />
+      <ProfileEditContainer />
     </ScreenContainer>
   );
 };
